add a scenario in sign service and rename event
