+++ src/lambda/trips/TripEvent.ts
